+++ grpc/grpc_test.go
@@ -181,9 +181,11 @@ func TestConfig_DefaultPortInError(t *testing.T) {
 
 type testLogger struct{ t *testing.T }
 
+func (l *testLogger) Debug(msg string, args ...any) { l.t.Log(append([]any{"DEBUG", msg}, args...)...) }
 func (l *testLogger) Info(msg string, args ...any) {
 	l.t.Log(append([]any{"INFO ", msg}, args...)...)
 }
+func (l *testLogger) Warn(msg string, args ...any)  { l.t.Log(append([]any{"WARN ", msg}, args...)...) }
 func (l *testLogger) Error(msg string, args ...any) {
 	l.t.Log(append([]any{"ERROR", msg}, args...)...)
 }