+++ grpcclient/grpcclient.go
@@ -55,6 +55,7 @@ import (
 
 // Logger is satisfied by [log/slog.Logger] and most structured loggers.
 type Logger interface {
+	Debug(msg string, args ...any)
 	Info(msg string, args ...any)
 	Warn(msg string, args ...any)
 	Error(msg string, args ...any)
@@ -62,6 +63,7 @@ type Logger interface {
 
 type nopLogger struct{}
 
+func (nopLogger) Debug(string, ...any) {}
 func (nopLogger) Info(string, ...any)  {}
 func (nopLogger) Warn(string, ...any)  {}
 func (nopLogger) Error(string, ...any) {}
@@ -272,7 +274,7 @@ func (c *Component) Stop(ctx context.Context) error {
 	select {
 	case <-done:
 	case <-ctx.Done():
-		c.log.Error("grpc-client: connection close timed out during shutdown")
+		c.log.Warn("grpc-client: connection close timed out during shutdown")
 	}
 	return nil
 }